logMalloc and logRealloc should use "size_t" instead of assuming it will be either 32 or 64 bits (it will change cased on whether we are using -m32 or not).

+++ defs.h
@@ -118,8 +118,8 @@ void* logLoadInst(Addr src_addr, UInt dest);
 void* logStoreInst(UInt src, Addr dest_addr); 
 void* logStoreInstConst(Addr dest_addr); 
 
-void logMalloc(Addr addr, UInt64 size);
-void logRealloc(Addr old_addr, Addr new_addr, UInt64 size);
+void logMalloc(Addr addr, size_t size);
+void logRealloc(Addr old_addr, Addr new_addr, size_t size);
 void logFree(Addr addr);
 
 void logPhiNode(UInt dest, UInt src, UInt num_cont_dep, ...);
